dasdf

+++ ProjectRoot/bash/calibration.sh
@@ -13,10 +13,10 @@ fi
 
 
 cent_node="32B1728"
-leaf_node_serials=("32B172B" "32C793E" "32B1708" "337D42D" "32C79F7" "32C7920" "32C79BE" "32C79C6" "32C7981")
+leaf_node_serials=("32B172B" "32C793E" "32B1708" "337D42D" "32C7920" "32C79BE" "32C79C6" "32C7981" "32C79F7")
 # leaf_node_serials=("32C79C6")
 
-remote_nodes=("rpi4m1@192.168.5.241" "rpi4m2@192.168.5.242" "rpi4m3@192.168.5.243" "rpi4m4@192.168.5.244" "rpi4m5@192.168.5.245" "rpi4m5@192.168.5.245" "rpi4compute@192.168.5.246" "nuc01@192.168.5.248" "nuc01@192.168.5.248")
+remote_nodes=("rpi4m1@192.168.5.241" "rpi4m2@192.168.5.242" "rpi4m3@192.168.5.243" "rpi4m4@192.168.5.244" "rpi4m5@192.168.5.245" "rpi4compute@192.168.5.246" "nuc01@192.168.5.248" "nuc01@192.168.5.248" "")
 # remote_nodes=("nuc01@192.168.5.248")
 
 cleanup() {
@@ -24,7 +24,11 @@ cleanup() {
     for (( j=0; j<${#remote_nodes[@]}; j++ ))
     do
         node_name="${remote_nodes[$i]}"
-        ssh ${node_name} "bash -c 'pkill -9 CA_calib'"
+        if [ "${node_name}" == "" ]; then
+            bash -c 'pkill -9 CA_'
+        else
+            ssh ${node_name} "bash -c 'pkill -9 CA_'"
+        fi
         sleep 1
     done
 }
@@ -42,15 +46,29 @@ do
     echo "Iteration $i for leaf node ${node_serial}: Starting..."
 
     # Start two commands in detached screen sessions
-    ssh ${node_name} "bash -c 'pkill -9 CA_calib'"
+    if [ "${node_name}" == "" ]; then
+        bash -c 'pkill -9 CA_'
+    else
+        ssh ${node_name} "bash -c 'pkill -9 CA_'"
+    fi
     sleep 1
+    
     if $CLEAN_SLATE; then
-        ssh ${node_name} "bash -c 'rm -rf \$HOME/OTA-C/ProjectRoot/storage/calibration/*'"
+        if [ "${node_name}" != "" ]; then
+            ssh ${node_name} "bash -c 'rm -rf \$HOME/OTA-C/ProjectRoot/storage/calibration/*'"
+        fi
         sleep 1
     fi
-    screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_leaf_${node_serial}.log" -S session_leaf_${node_serial} bash -c "ssh ${node_name} 'cd \$HOME/OTA-C/ProjectRoot/build && ./CA_calib leaf ${node_serial} ${cent_node}'"
-    sleep 1
-    screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_cent_${node_serial}.log" -S session_cent_${node_serial} bash -c "cd $HOME/OTA-C/ProjectRoot/build/ && ./CA_calib cent ${cent_node} ${node_serial}"
+
+    if [ "${node_name}" == "" ]; then
+        screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_leaf_${node_serial}.log" -S session_leaf_${node_serial} bash -c "cd $HOME/OTA-C/ProjectRoot/build && ./CA_calib leaf ${node_serial} ${cent_node}"
+        sleep 1
+        screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_cent_${node_serial}.log" -S session_cent_${node_serial} bash -c "cd $HOME/OTA-C/ProjectRoot/build/ && ./CA_calib cent ${cent_node} ${node_serial}"
+    else
+        screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_leaf_${node_serial}.log" -S session_leaf_${node_serial} bash -c "ssh ${node_name} 'cd \$HOME/OTA-C/ProjectRoot/build && ./CA_calib leaf ${node_serial} ${cent_node}'"
+        sleep 1
+        screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_cent_${node_serial}.log" -S session_cent_${node_serial} bash -c "cd $HOME/OTA-C/ProjectRoot/build/ && ./CA_calib cent ${cent_node} ${node_serial}"
+    fi
 
     start_time=$(date +%s)
     while true; do
